SpaceWarpPatcher reorganized (into the first Galactic Empire)
