+++ src/SpaceWarp.Core/Properties/AssemblyInfo.cs
