+++ src/SpaceWarpPatcher/Patches/ResourceFlowPatch.cs
@@ -5,19 +5,19 @@ using Mono.Collections.Generic;
 using MonoMod.Cil;
 using MonoMod.Utils;
 
-namespace SpaceWarpPatcher;
+namespace SpaceWarpPatcher.Patches;
 
 /// <summary>
 /// Patcher for the game's main DLL
 /// </summary>
 [UsedImplicitly]
-public class AssemblyCSharpPatcher
+public class ResourceFlowPatch
 {
     /// <summary>
     /// The target DLLs to patch
     /// </summary>
     [UsedImplicitly]
-    public static IEnumerable<string> TargetDLLs => new[] { "Assembly-CSharp.dll"};
+    public static IEnumerable<string> TargetDLLs => new[] { "Assembly-CSharp.dll" };
 
     /// <summary>
     /// Patches the target DLL
@@ -26,42 +26,31 @@ public class AssemblyCSharpPatcher
     [UsedImplicitly]
     public static void Patch(ref AssemblyDefinition assemblyDefinition)
     {
-        // AssemblyDefinition coreAssembly = null;
-        // var dir = new DirectoryInfo(Paths.PluginPath);
-        // foreach (var file in dir.EnumerateFiles("SpaceWarp.Core.dll", SearchOption.AllDirectories))
-        // {
-        //     coreAssembly = AssemblyDefinition.ReadAssembly(file.FullName);
-        // }
-        //
-        // if (coreAssembly == null)
-        // {
-        //     throw new Exception("Could not find SpaceWarp Core");
-        // }
-
         var firstTargetType = assemblyDefinition.MainModule.Types.First(t => t.Name == "PartOwnerComponent");
         var boolType = firstTargetType.Fields
-            .Select(x => x.FieldType).First(x => x.MetadataType == MetadataType.Boolean)!;
-        firstTargetType.Fields.Add(new FieldDefinition("HasRegisteredPartComponentsForFixedUpdate",FieldAttributes.Public,boolType));
+            .Select(x => x.FieldType)
+            .First(x => x.MetadataType == MetadataType.Boolean)!;
+        firstTargetType.Fields.Add(
+            new FieldDefinition("HasRegisteredPartComponentsForFixedUpdate", FieldAttributes.Public, boolType)
+        );
         var field = firstTargetType.Fields.First(x => x.Name == "HasRegisteredPartComponentsForFixedUpdate");
-        // Now later we harmony patch the initializer for partownercomponent
+        // Now later we harmony patch the initializer for PartOwnerComponent
 
         var targetMethod = firstTargetType.Methods.First(method => method.Name == "OnFixedUpdate");
-        var methodCallA = assemblyDefinition.MainModule.Types.First(t => t.Name == "ResourceFlowRequestManager").Methods
+        var methodCallA = assemblyDefinition.MainModule.Types
+            .First(t => t.Name == "ResourceFlowRequestManager")
+            .Methods
             .First(m => m.Name == "UpdateFlowRequests");
         var methodCallB = firstTargetType.Properties
-            .First(definition => definition.Name == "ResourceFlowRequestManager").GetMethod;
-        // var context = new ILContext(targetMethod);
-        // var cursor = new ILCursor(context);
-        // var elseEndLabel  = cursor.DefineLabel();
-        // var elseBeginLabel = cursor.DefineLabel();
-        // cursor.GotoNext(MoveType.Before)
-        // cursor.GotoNext(MoveType.After, instruction => instruction.MatchCallOrCallvirt(methodCallA));
-        var insts = targetMethod.Body.Instructions;
-        Collection<Instruction> newInstructions = new Collection<Instruction>();
+            .First(definition => definition.Name == "ResourceFlowRequestManager")
+            .GetMethod;
+
+        var instructions = targetMethod.Body.Instructions;
+        var newInstructions = new Collection<Instruction>();
         var nextIsTarget = false;
         Instruction done = null;
         Instruction @else = null;
-        foreach (var currentInstruction in insts)
+        foreach (var currentInstruction in instructions)
         {
             if (nextIsTarget)
             {
@@ -115,8 +104,7 @@ public class AssemblyCSharpPatcher
             }
         }
 
-
-        insts.Clear();
-        insts.AddRange(newInstructions);
+        instructions.Clear();
+        instructions.AddRange(newInstructions);
     }
 }
