+++ src/SpaceWarpPatcher/Properties/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("SpaceWarp.Core")]
+[assembly: InternalsVisibleTo("SpaceWarp.UI")]
